+++ src/components/item/ItemDetails.js
@@ -1,12 +1,13 @@
 import React from "react";
-import { Container, Button } from "reactstrap";
+import { Container } from "reactstrap";
 import moment from "moment";
+import ContactButton from "./ContactButton.js";
 
 // Detailed view of status (displayed below MapView)
 const ItemDetails = ({ userSignedIn, item }) => {
   if (item) {
     return (
-      <Container style={{ marginTop: 10 }}>
+      <Container style={{ marginTop: 30 }}>
         {/* Title */}
         <h5>{item.title} </h5>
         <div className="lead">{item.location.description}</div>
@@ -14,19 +15,7 @@ const ItemDetails = ({ userSignedIn, item }) => {
           <span style={{ fontSize: 14, color: "gray" }}>
             Date Posted: {moment(item.createdAt.toDate()).calendar()}
           </span>
-
-          {/* Check if user ID is the same as post author ID */}
-          <Button
-            onClick={() => {
-              item.authorId === userSignedIn
-                ? alert("You cannot contact yourself!")
-                : alert("Coming soon!");
-            }}
-            color="info"
-            style={{ float: "right" }}
-          >
-            Contact
-          </Button>
+          <ContactButton item={item} userSignedIn={userSignedIn} />
         </div>
         <hr className="my-3" />
         <p>{item.description}</p>
